requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
autopep8==2.0.2
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.1
django-cors-headers==4.1.0
django-defender==0.9.7
django-ses==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==7.0.0
gunicorn==21.2.0
idna==3.4
isort==5.13.2
jmespath==1.0.1
mando==0.7.1
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
radon==6.0.1
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
xenon==0.9.1